Separate realized instruments from generic composable traits
