+++ src/main/scala/org/retistruen/package.scala
@@ -29,9 +29,11 @@ trait Emitter[T] extends Named {
   private def register(rec: Receiver[T]): Unit =
     receivers = receivers + rec
 
-  def -->(others: Receiver[T]*): Unit = others.foreach(register(_))
+  def >>>(others: Receiver[T]*): Unit = others.foreach(register(_))
 
-  def -->(other: Receiver[T] with Emitter[T]): Emitter[T] = {
+  def >>(other: Receiver[T]): Unit = register(other)
+
+  def >>(other: Receiver[T] with Emitter[T]): Emitter[T] = {
     register(other)
     other
   }
@@ -50,24 +52,17 @@ trait Emitter[T] extends Named {
 
 }
 
-/**
- * Es un Receiver que maneja un buffer de los últimos "size" Datums recibidos
- */
 trait BufferedReceiver[T] extends Receiver[T] {
 
   val size: Int
 
-  private var data: Queue[Datum[T]] = Queue.empty
+  private var data: Seq[Datum[T]] = Seq.empty
 
   def buffer: Seq[Datum[T]] = data
 
-  override def receive(emitter: Emitter[T], datum: Datum[T]) = {
-    if (data.size >= size) {
-      val (_, queue) = data.dequeue
-      data = queue.enqueue(datum)
-    } else {
-      data = data.enqueue(datum)
-    }
+  def receive(emitter: Emitter[T], datum: Datum[T]) = {
+    data = data :+ datum
+    data = data.drop(data.size - size)
   }
 
 }
@@ -103,79 +98,5 @@ trait BufferedFunctor[T, R] extends Functor[T, R] with CachedEmitter[R] with Buf
     super.receive(emitter, datum)
     emit(result(buffer))
   }
-
-}
-
-class Variable[T](val name: String) extends UnbufferedFunctor[T, T] {
-
-  protected def result(datum: Datum[T]) = datum
-
-}
-
-class BufferedMax[N: Ordering](val name: String, val size: Int) extends BufferedFunctor[N, N] {
-
-  def result(data: Seq[Datum[N]]) = Datum(data.map(_.value).max)
-
-}
-
-class Max[N: Ordering](val name: String) extends UnbufferedFunctor[N, N] {
-
-  val o = implicitly[Ordering[N]]
-  import o._
-
-  def result(received: Datum[N]): Datum[N] = last match {
-    case Some(previous) ⇒ if (received.value > previous.value) received.fresh else previous.fresh
-    case None           ⇒ received.fresh
-  }
-
-}
-
-class RecordingReceiver[T](val name: String, val capacity: Option[Int] = None) extends Receiver[T] {
-
-  private var buffer: Seq[Datum[T]] = Seq.empty
-
-  def data = buffer
-
-  def receive(emitter: Emitter[T], datum: Datum[T]) = {
-    buffer = buffer :+ datum
-    for (c ← capacity) buffer = buffer.drop(buffer.size - c)
-  }
-
-  def clear =
-    buffer = Seq.empty
-
-}
-
-class SourceEmitter[T](val name: String) extends CachedEmitter[T]
-
-class Model {
-
-  def source[T](name: String): Emitter[T] = new SourceEmitter[T](name)
-
-  trait Builder[T] {
-    val source: Emitter[T]
-    protected def name(tag: String): String =
-      "%s.%s" format (source.name, tag)
-  }
-
-  class FactoryBuilder[T](val source: Emitter[T]) extends Builder[T] {
-    def |->(): Emitter[T] = source
-  }
-
-  protected implicit def sourceToFactoryBuilder[T: Ordering](e: Emitter[T]): FactoryBuilder[T] = new FactoryBuilder[T](e)
-
-  class MaxBuilder[T: Ordering](val source: Emitter[T]) extends Builder[T] {
-    def max(size: Int) = new BufferedMax[T](name("max" + size), size)
-    def max = new Max[T](name("max"))
-  }
-
-  protected implicit def sourceToMaxBuilder[T: Ordering](e: Emitter[T]): MaxBuilder[T] = new MaxBuilder[T](e)
-
-}
-
-class Sensors extends Model {
-
-  val s1 = source[Int]("s1")
-  val s2 = source[Int]("s2")
 
 }
